gp_score parseInt -> parseFloat for pdf generation

+++ src/app/html2pdf/page.tsx
@@ -200,7 +200,7 @@ const Page = () => {
           <div className="pb-6">
             <div id="gp-score" className="bg-[#542C06] text-white p-4 text-3xl font-bold rounded-2xl w-56 flex justify-center">
               <span>
-                GP Score- {gpScore && parseInt(gpScore)}
+                GP Score- {gpScore && parseFloat(gpScore)}
               </span>
             </div>
           </div>
